modification done in code to make it working

+++ src/main/java/com/example/loginApp/repository/PermissionRepo.java
@@ -4,6 +4,9 @@ import com.example.loginApp.model.PermissionVO;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface PermissionRepo extends JpaRepository<PermissionVO, Long> {
+public interface PermissionRepo extends JpaRepository<PermissionVO, Integer> {
+
 }
