+++ src/main/java/com/example/loginApp/repository/RoleRepo.java
@@ -5,6 +5,10 @@ import com.example.loginApp.model.RoleVO;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface RoleRepo extends JpaRepository<RoleVO, Long> {
+public interface RoleRepo extends JpaRepository<RoleVO, Integer> {
+
+
 }
